+++ src/body/Stash Batch Save.user.js
@@ -1,4 +1,4 @@
-(function() {
+(function () {
     'use strict';
 
     const {
@@ -8,9 +8,16 @@
         waitForElementClass,
         waitForElementByXpath,
         getElementByXpath,
+        getElementsByXpath,
         getClosestAncestor,
         sortElementChildren,
+        createElementFromHTML,
     } = unsafeWindow.stash;
+
+    document.body.appendChild(document.createElement('style')).textContent = `
+    .search-item > div.row:first-child > div.col-md-6.my-1 > div:first-child { display: flex; flex-direction: column; }
+    .tagger-remove { order: 10; }
+    `;
 
     let running = false;
     const buttons = [];
@@ -23,6 +30,13 @@
         stash.setProgress((maxCount - buttons.length) / maxCount * 100);
         if (button) {
             const searchItem = getClosestAncestor(button, '.search-item');
+            if (searchItem.classList.contains('d-none')) {
+                setTimeout(() => {
+                    run();
+                }, 0);
+                return;
+            }
+
             const { id } = stash.parseSearchItem(searchItem);
             sceneId = id;
             if (!button.disabled) {
@@ -41,7 +55,7 @@
         if (running && evt.detail.data?.sceneUpdate?.id === sceneId) {
             setTimeout(() => {
                 run();
-            }, 0)
+            }, 0);
         }
     }
 
@@ -107,4 +121,39 @@
 
     stash.addEventListener('tagger:mutations:searchitems', checkSaveButtonDisplay);
 
+    async function initRemoveButtons() {
+        const nodes = getElementsByXpath("//button[contains(@class, 'btn-primary') and text()='Scrape by fragment']");
+        const buttons = [];
+        let node = null;
+        while (node = nodes.iterateNext()) {
+            buttons.push(node);
+        }
+        for (const button of buttons) {
+            const searchItem = getClosestAncestor(button, '.search-item');
+
+            const removeButtonExists = searchItem.querySelector('.tagger-remove');
+            if (removeButtonExists) {
+                continue;
+            }
+
+            const removeEl = createElementFromHTML('<div class="mt-2 text-right tagger-remove"><button class="btn btn-danger">Remove</button></div>');
+            const removeButton = removeEl.querySelector('button');
+            button.parentElement.parentElement.appendChild(removeEl);
+            removeButton.addEventListener('click', async () => {
+                searchItem.classList.add('d-none');
+            });
+        }
+    }
+
+    stash.addEventListener('page:studio:scenes', function () {
+        waitForElementByXpath("//button[contains(@class, 'btn-primary') and text()='Scrape by fragment']", initRemoveButtons);
+    });
+
+    stash.addEventListener('page:performer:scenes', function () {
+        waitForElementByXpath("//button[contains(@class, 'btn-primary') and text()='Scrape by fragment']", initRemoveButtons);
+    });
+
+    stash.addEventListener('page:scenes', function () {
+        waitForElementByXpath("//button[contains(@class, 'btn-primary') and text()='Scrape by fragment']", initRemoveButtons);
+    });
 })();