+++ src/header/Stash Markers Autoscroll.user.js
@@ -1,8 +1,8 @@
 // ==UserScript==
-// @name        Stash Studio Image And Parent On Create
+// @name        Stash Markers Autoscroll
 // @namespace   %NAMESPACE%
-// @description Set studio image and parent when creating from StashDB. Requires userscript_functions stash plugin
-// @version     0.3.0
+// @description Automatically scrolls markers page
+// @version     0.1.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow