+++ CHANGELOG.md
@@ -1,5 +1,58 @@
 # Changelog
 
+## 0.24.2 Stash Userscripts Bundle
+### 0.5.3 Stash Batch Save
+#### Added
+- Add remove button to prevent saving
+#### Changed
+- Add css to set order of submit draft and batch save remove buttons
+### 0.1.1 Stash Scene Tagger Draft Submit
+#### Changed
+- Add css to set order of submit draft and batch save remove buttons
+
+## 0.24.1 Stash Userscripts Bundle
+### 0.3.1 Stash Stats
+#### Fixed
+- Fix showing unmatched instead of matched stats
+
+## 0.24.0 Stash Userscripts Bundle
+### 0.6.0 Stash Batch Result Toggle
+#### Changed
+- Added director and studio code toggling
+### 0.7.0 Stash Scene Tagger Colorizer
+#### Changed
+- Added director and studio code coloring
+
+## 0.23.0 Stash Userscripts Bundle
+### 0.1.0 Stash Markers Autoscroll
+#### Added
+- Add userscript
+### 0.6.0 Stash Scene Tagger Colorizer
+#### Changed
+- Added tag coloring
+
+## 0.6.0 Userscript Functions Plugin
+### Fixed
+- Updated plugin to work with Stash v0.24.3
+
+## 0.22.0 Stash Userscripts Bundle
+#### Fixed
+- Updated userscripts to work with Stash v0.24.3
+### 0.5.0 Stash Batch Result Toggle
+### 0.2.0 Stash Markdown
+### 0.3.0 Stash Performer Image Cropper
+### 0.5.0 Stash Scene Tagger Colorizer
+### 0.3.0 Stash Set Stashbox Favorite Performers
+### 0.2.0 Stash StashID Icon
+### 0.5.0 Stash StashID Input
+### 0.3.0 Stash Stats
+### 0.2.0 Stash Tag Image Cropper
+
+#### Removed
+- Removed obsolete userscripts
+### Stash Scene Tagger Linkify
+### Stash Studio Image And Parent On Create
+
 ## 0.21.1 Stash Userscripts Bundle
 ### 0.1.1 Stash Markdown
 #### Fixed